+++ colonyShip_test.go
@@ -9,12 +9,12 @@ import (
 func TestColonyShip_GetSpeed(t *testing.T) {
 	cs := newColonyShip()
 	speed := cs.GetSpeed(Researches{ImpulseDrive: 6})
-	assert.Equal(t, 5500, speed)
+	assert.Equal(t, int64(5500), speed)
 
 }
 
 func TestColony_GetCargoCapacity(t *testing.T) {
 	cs := newColonyShip()
-	assert.Equal(t, 10500, cs.GetCargoCapacity(Researches{HyperspaceTechnology: 8}, false))
+	assert.Equal(t, int64(10500), cs.GetCargoCapacity(Researches{HyperspaceTechnology: 8}, false))
 
 }
