+++ interfaces.go
@@ -43,12 +43,12 @@ type Wrapper interface {
 	IsConnected() bool
 	GetUsername() string
 	GetUniverseName() string
-	GetUniverseSpeed() int
-	GetUniverseSpeedFleet() int
-	GetResearchSpeed() int
-	SetResearchSpeed(int)
-	GetNbSystems() int
-	SetNbSystems(int)
+	GetUniverseSpeed() int64
+	GetUniverseSpeedFleet() int64
+	GetResearchSpeed() int64
+	SetResearchSpeed(int64)
+	GetNbSystems() int64
+	SetNbSystems(int64)
 	IsDonutGalaxy() bool
 	IsDonutSystem() bool
 	FleetDeutSaveFactor() float64
@@ -57,15 +57,15 @@ type Wrapper interface {
 	Location() *time.Location
 	IsUnderAttack() (bool, error)
 	GetUserInfos() UserInfos
-	SendMessage(playerID int, message string) error
-	SendMessageAlliance(associationID int, message string) error
+	SendMessage(playerID int64, message string) error
+	SendMessageAlliance(associationID int64, message string) error
 	ReconnectChat() bool
 	GetFleets() ([]Fleet, Slots)
 	GetFleetsFromEventList() []Fleet
 	CancelFleet(FleetID) error
 	GetAttacks() ([]AttackEvent, error)
 	GetAttacksUsing(CelestialID) ([]AttackEvent, error)
-	GalaxyInfos(galaxy, system int) (SystemInfos, error)
+	GalaxyInfos(galaxy, system int64) (SystemInfos, error)
 	GetCachedResearch() Researches
 	GetResearch() Researches
 	GetCachedPlanets() []Planet
@@ -84,12 +84,12 @@ type Wrapper interface {
 	Abandon(interface{}) error
 	GetEspionageReportMessages() ([]EspionageReportSummary, error)
 	GetEspionageReportFor(Coordinate) (EspionageReport, error)
-	GetEspionageReport(msgID int) (EspionageReport, error)
+	GetEspionageReport(msgID int64) (EspionageReport, error)
 	GetCombatReportSummaryFor(Coordinate) (CombatReportSummary, error)
 	//GetCombatReport(msgID int) (CombatReport, error)
-	DeleteMessage(msgID int) error
-	Distance(origin, destination Coordinate) int
-	FlightTime(origin, destination Coordinate, speed Speed, ships ShipsInfos) (secs, fuel int)
+	DeleteMessage(msgID int64) error
+	Distance(origin, destination Coordinate) int64
+	FlightTime(origin, destination Coordinate, speed Speed, ships ShipsInfos) (secs, fuel int64)
 	RegisterChatCallback(func(ChatMsg))
 	RegisterAuctioneerCallback(func([]byte))
 	RegisterHTMLInterceptor(func(method, url string, params, payload url.Values, pageHTML []byte))
@@ -97,22 +97,22 @@ type Wrapper interface {
 	BuyOfferOfTheDay() error
 	BytesDownloaded() int64
 	BytesUploaded() int64
-	CreateUnion(fleet Fleet) (int, error)
+	CreateUnion(fleet Fleet) (int64, error)
 
 	// Planet or Moon functions
 	GetResources(CelestialID) (Resources, error)
 	GetResourcesDetails(CelestialID) (ResourcesDetails, error)
-	SendFleet(celestialID CelestialID, ships []Quantifiable, speed Speed, where Coordinate, mission MissionID, resources Resources, expeditiontime, unionID int) (Fleet, error)
-	EnsureFleet(celestialID CelestialID, ships []Quantifiable, speed Speed, where Coordinate, mission MissionID, resources Resources, expeditiontime, unionID int) (Fleet, error)
-	Build(celestialID CelestialID, id ID, nbr int) error
+	SendFleet(celestialID CelestialID, ships []Quantifiable, speed Speed, where Coordinate, mission MissionID, resources Resources, expeditiontime, unionID int64) (Fleet, error)
+	EnsureFleet(celestialID CelestialID, ships []Quantifiable, speed Speed, where Coordinate, mission MissionID, resources Resources, expeditiontime, unionID int64) (Fleet, error)
+	Build(celestialID CelestialID, id ID, nbr int64) error
 	BuildCancelable(CelestialID, ID) error
-	BuildProduction(celestialID CelestialID, id ID, nbr int) error
+	BuildProduction(celestialID CelestialID, id ID, nbr int64) error
 	BuildBuilding(celestialID CelestialID, buildingID ID) error
-	BuildDefense(celestialID CelestialID, defenseID ID, nbr int) error
-	BuildShips(celestialID CelestialID, shipID ID, nbr int) error
+	BuildDefense(celestialID CelestialID, defenseID ID, nbr int64) error
+	BuildShips(celestialID CelestialID, shipID ID, nbr int64) error
 	CancelBuilding(CelestialID) error
 	TearDown(celestialID CelestialID, id ID) error
-	ConstructionsBeingBuilt(CelestialID) (buildingID ID, buildingCountdown int, researchID ID, researchCountdown int)
+	ConstructionsBeingBuilt(CelestialID) (buildingID ID, buildingCountdown int64, researchID ID, researchCountdown int64)
 	GetProduction(CelestialID) ([]Quantifiable, error)
 	GetFacilities(CelestialID) (Facilities, error)
 	GetDefense(CelestialID) (DefensesInfos, error)
@@ -124,7 +124,7 @@ type Wrapper interface {
 	// Planet specific functions
 	GetResourceSettings(PlanetID) (ResourceSettings, error)
 	SetResourceSettings(PlanetID, ResourceSettings) error
-	SendIPM(PlanetID, Coordinate, int, ID) (int, error)
+	SendIPM(PlanetID, Coordinate, int64, ID) (int64, error)
 	//GetResourcesProductionRatio(PlanetID) (float64, error)
 	GetResourcesProductions(PlanetID) (Resources, error)
 	GetResourcesProductionsLight(ResourcesBuildings, Researches, ResourceSettings, Temperature) Resources
@@ -139,16 +139,16 @@ type Wrapper interface {
 type BaseOgameObj interface {
 	GetID() ID
 	GetName() string
-	ConstructionTime(nbr, universeSpeed int, facilities Facilities) time.Duration
-	GetRequirements() map[ID]int
-	GetPrice(int) Resources
-	IsAvailable(CelestialType, LazyResourcesBuildings, LazyFacilities, LazyResearches, int) bool
+	ConstructionTime(nbr, universeSpeed int64, facilities Facilities) time.Duration
+	GetRequirements() map[ID]int64
+	GetPrice(int64) Resources
+	IsAvailable(CelestialType, LazyResourcesBuildings, LazyFacilities, LazyResearches, int64) bool
 }
 
 // Levelable base interface for all levelable ogame objects (buildings, technologies)
 type Levelable interface {
 	BaseOgameObj
-	GetLevel(LazyResourcesBuildings, LazyFacilities, LazyResearches) int
+	GetLevel(LazyResourcesBuildings, LazyFacilities, LazyResearches) int64
 }
 
 // Technology interface that all technologies implement
@@ -164,19 +164,19 @@ type Building interface {
 // DefenderObj base interface for all defensive units (ships, defenses)
 type DefenderObj interface {
 	BaseOgameObj
-	GetStructuralIntegrity(Researches) int
-	GetShieldPower(Researches) int
-	GetWeaponPower(Researches) int
-	GetRapidfireFrom() map[ID]int
+	GetStructuralIntegrity(Researches) int64
+	GetShieldPower(Researches) int64
+	GetWeaponPower(Researches) int64
+	GetRapidfireFrom() map[ID]int64
 }
 
 // Ship interface implemented by all ships units
 type Ship interface {
 	DefenderObj
-	GetCargoCapacity(techs Researches, probeRaids bool) int
-	GetSpeed(Researches) int
-	GetFuelConsumption() int
-	GetRapidfireAgainst() map[ID]int
+	GetCargoCapacity(techs Researches, probeRaids bool) int64
+	GetSpeed(Researches) int64
+	GetFuelConsumption() int64
+	GetRapidfireAgainst() map[ID]int64
 }
 
 // Defense interface implemented by all defenses units
@@ -194,15 +194,15 @@ type Celestial interface {
 	GetResources() (Resources, error)
 	GetResourcesDetails() (ResourcesDetails, error)
 	GetFacilities() (Facilities, error)
-	SendFleet([]Quantifiable, Speed, Coordinate, MissionID, Resources, int, int) (Fleet, error)
-	EnsureFleet([]Quantifiable, Speed, Coordinate, MissionID, Resources, int, int) (Fleet, error)
+	SendFleet([]Quantifiable, Speed, Coordinate, MissionID, Resources, int64, int64) (Fleet, error)
+	EnsureFleet([]Quantifiable, Speed, Coordinate, MissionID, Resources, int64, int64) (Fleet, error)
 	GetDefense() (DefensesInfos, error)
 	GetShips() (ShipsInfos, error)
-	BuildDefense(defenseID ID, nbr int) error
-	ConstructionsBeingBuilt() (ID, int, ID, int)
+	BuildDefense(defenseID ID, nbr int64) error
+	ConstructionsBeingBuilt() (ID, int64, ID, int64)
 	GetProduction() ([]Quantifiable, error)
 	GetResourcesBuildings() (ResourcesBuildings, error)
-	Build(id ID, nbr int) error
+	Build(id ID, nbr int64) error
 	BuildBuilding(buildingID ID) error
 	BuildTechnology(technologyID ID) error
 	CancelResearch() error
