+++ reaper.go
@@ -14,9 +14,9 @@ func newReaper() *reaper {
 	r.BaseCargoCapacity = 10000
 	r.BaseSpeed = 7000
 	r.FuelConsumption = 900
-	r.RapidfireFrom = map[ID]int{DeathstarID: 10, IonCannonID: 2}
-	r.RapidfireAgainst = map[ID]int{EspionageProbeID: 5, SolarSatelliteID: 5, CrawlerID: 5, BattleshipID: 7, BattlecruiserID: 7, BomberID: 4, DestroyerID: 3}
+	r.RapidfireFrom = map[ID]int64{DeathstarID: 10, IonCannonID: 2}
+	r.RapidfireAgainst = map[ID]int64{EspionageProbeID: 5, SolarSatelliteID: 5, CrawlerID: 5, BattleshipID: 7, BattlecruiserID: 7, BomberID: 4, DestroyerID: 3}
 	r.Price = Resources{Metal: 85000, Crystal: 55000, Deuterium: 20000}
-	r.Requirements = map[ID]int{ShipyardID: 10, HyperspaceTechnologyID: 6, HyperspaceDriveID: 7, ShieldingTechnologyID: 6}
+	r.Requirements = map[ID]int64{ShipyardID: 10, HyperspaceTechnologyID: 6, HyperspaceDriveID: 7, ShieldingTechnologyID: 6}
 	return r
 }
